requirements.txt
langchain==0.1.0
chromadb==0.4.18
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai==1.3.7
fastapi==0.104.1
pydantic==2.5.2
pydantic-settings==2.1.0
streamlit==1.29.0
python-dotenv==1.0.0
uvicorn==0.24.0.post1
pytest==7.4.3
httpx==0.25.2
aiohttp==3.9.1
tiktoken==0.5.1
tenacity==8.2.3
redis==4.5.4
pypdf2==3.0.1
python-docx==0.8.11